+++ src/lang-xslt/main/uk/co/reecedunn/intellij/plugin/xpath/completion/XPathCompletionContributor.kt
@@ -16,10 +16,12 @@
 package uk.co.reecedunn.intellij.plugin.xpath.completion
 
 import uk.co.reecedunn.intellij.plugin.core.completion.CompletionContributorEx
+import uk.co.reecedunn.intellij.plugin.xpath.completion.filters.XPathForwardOrReverseAxisFilter
 import uk.co.reecedunn.intellij.plugin.xpath.completion.filters.XPathItemTypeFilter
 import uk.co.reecedunn.intellij.plugin.xpath.completion.filters.XPathKindTestFilter
 import uk.co.reecedunn.intellij.plugin.xpath.completion.filters.XPathSequenceTypeFilter
 import uk.co.reecedunn.intellij.plugin.xpath.completion.property.XPathVersion
+import uk.co.reecedunn.intellij.plugin.xpath.completion.providers.XPathForwardOrReverseAxisProvider
 import uk.co.reecedunn.intellij.plugin.xpath.completion.providers.XPathItemTypeProvider
 import uk.co.reecedunn.intellij.plugin.xpath.completion.providers.XPathKindTestProvider
 import uk.co.reecedunn.intellij.plugin.xpath.completion.providers.XPathSequenceTypeProvider
@@ -27,6 +29,9 @@ import uk.co.reecedunn.intellij.plugin.xslt.psi.isIntellijXPathPluginEnabled
 
 class XPathCompletionContributor : CompletionContributorEx() {
     private fun registerXPathCompletionProviders() {
+        // XPath 3.1 EBNF (41) ForwardAxis ; XPath 3.1 EBNF (44) ReverseAxis
+        builder().withFilter(XPathForwardOrReverseAxisFilter).addCompletions(XPathForwardOrReverseAxisProvider)
+
         // XPath 3.1 EBNF (79) SequenceType
         builder().withFilter(XPathSequenceTypeFilter).withProperty(XPathVersion)
             .addCompletions(XPathSequenceTypeProvider)