+++ CHANGELOG.md
@@ -14,6 +14,7 @@ IntelliJ Integration:
 
 Code Completion:
 
+1.  Keyword completion support for `ForwardAxis` and `ReverseAxis`.
 1.  Keyword completion support for `KindTest` based `NodeTest`s and `ItemType`s.
 1.  Keyword completion support for `SequenceType`s.